app/requirements.txt
Flask==1.1.2
Jinja2==2.11.3
mysql-connector-python==8.0.25
Werkzeug==1.0.1
MarkupSafe==1.1.1
itsdangerous==1.1.0